Redo how we do system mount discovery
... and how we propagate these mounts to the mount farm. In particular,
if we mark a directory as hidden, it's not enough to not copy it over to the
mount farm - if we just do that, _add_missing_children may later re-add
these directories as bind mounts anyway. Instead, we must propagate
hidden nodes all the way to the mount farm, and just ignore them at the
final stage when performing the actual mounting.

Signed-off-by: Olaf Kirch <okir@suse.com>