Work around issues with overlayfs using mounts with children
The problem occurs when trying to overlay something over /usr while there's
a separate /usr/local mount.

When we change the file system view to private, all mounts are marked as MNT_LOCKED,
and we're not permitted to create an overlay mount when the lowerdir has children
that are locked in this way.

To work around this, we push down the overlays to a level where they don't conflict
with the child mount.

Signed-off-by: Olaf Kirch <okir@suse.de>